requirements.txt
Django==5.2.6
django-debug-toolbar==6.0.0
django-debug-toolbar-force==0.2
djangorestframework==3.16.1
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
celery[redis]==5.5.3
flower==2.0.1
djoser==2.3.3
django-filter
